ボタンを押した時にも通信が発生し、それをビューに返すようにした
