+++ NyanNyanEngine/HomeTimelineViewModel.swift
@@ -9,10 +9,11 @@
 import Foundation
 import RxSwift
 import RxCocoa
+import Alamofire
 
 protocol HomeTimelineViewModelInput: AnyObject {
     //TODO: 後々、日付型っぽいやつにする
-    var refreshExecutedAt: AnyObserver<String> { get }
+    var refreshExecutedAt: AnyObserver<String>? { get }
 }
 
 protocol HomeTimelineViewModelOutput: AnyObject {
@@ -20,17 +21,42 @@ protocol HomeTimelineViewModelOutput: AnyObject {
 }
 
 final class HomeTimelineViewModel: HomeTimelineViewModelInput, HomeTimelineViewModelOutput {
-    let refreshExecutedAt: AnyObserver<String>
+    private let disposeBag = DisposeBag()
+    
+    var refreshExecutedAt: AnyObserver<String>? = nil
     let statuses: Observable<[Status]?>
     
     init() {
-        let _refreshExecutedAt = PublishRelay<String>()
-        self.refreshExecutedAt = AnyObserver<String>() { event in
-            guard let text = event.element else { return }
-            _refreshExecutedAt.accept(text)
-        }
-        
-        let _statuses = BehaviorRelay<[Status]?>(value: nil)
+        let _statuses = BehaviorSubject<[Status]?>(value: nil)
         self.statuses = _statuses.asObservable()
+        
+        self.refreshExecutedAt = AnyObserver<String>() { updatedAt in
+            self.getResponse(url: "https://nyannyanengine-ios-d.firebaseapp.com/1.1/statuses/home_timeline.json")
+                .map { [unowned self] in self.toResponseBody(dataResponse: $0) }
+                .map { [unowned self] in self.toStatuses(data: $0) }
+                .bind(to: _statuses)
+                .disposed(by: self.disposeBag)
+            print(updatedAt.element)
+        }
     }
+    
+    func getResponse(url: String) -> Observable<DataResponse<String>> {
+        return Observable<DataResponse<String>>.create { observer in
+            Alamofire.request(url, method: .get)
+                .responseString(encoding: .utf8) { observer.onNext($0) }
+            return Disposables.create()
+        }
+    }
+    
+    private func toResponseBody(dataResponse: DataResponse<String>) -> Data? {
+        return dataResponse.value?.data(using: .utf8)
+    }
+    
+    private func toStatuses(data: Data?) -> [Status]? {
+        let decoder = JSONDecoder()
+        decoder.keyDecodingStrategy = .convertFromSnakeCase
+        guard let d = data else { return nil }
+        return try? decoder.decode([Status].self, from: d)
+    }
+    
 }
